Program 11, to find BB
A block of data is stored in memory locations from C071. Length of the
block is stored at C070. This is a program that searches for the first
occurrence of the data byte BB in the given block and the address of
this block in the HL pair. If the number is not found then the HL pair
should contain AAAA.